re architect structure final 2
